instead of using the hs cli, use applescript
the hs cli has a big downside, which is that when the configuration file has an error, then you can no longer communicate with hammerspoon because it failed to start and therefore to load the required ipc module.

instead we now require that user allows applescript by having this in its file `hs.allowAppleScript(true)`

with this we communicate with hammerspoon using applescript.